EA-162 WIP

+++ mobile-web/src/main/java/net/syscon/elite/persistence/LocationRepository.java
@@ -1,14 +1,15 @@
 package net.syscon.elite.persistence;
 
 
-import net.syscon.elite.web.api.model.Location;
-import net.syscon.elite.web.api.resource.LocationsResource.Order;
+
+import net.syscon.elite.v2.api.model.Location;
+import net.syscon.elite.v2.api.support.Order;
 
 import java.util.List;
 import java.util.Optional;
 
 public interface LocationRepository {
-	Optional<Location> findLocation(Long locationId);
-	List<Location> findLocations(String query, String orderByField, Order order, int offset, int limit);
-	List<Location> findLocationsByAgencyId(String caseLoadId, String agencyId, String query, int offset, int limit, String orderByField, Order order);
+	Optional<Location> findLocation(long locationId);
+	List<Location> findLocations(String query, String orderByField, Order order, long offset, long limit);
+	List<Location> findLocationsByAgencyId(String caseLoadId, String agencyId, String query, long offset, long limit, String orderByField, Order order);
 }
